# README-swagger.yaml
openapi: 3.0.3

info:
  title: Horse Training Planner API
  version: 1.0.0

servers:
  - url: http://localhost:3000/api

# GLOBAL JWT SECURITY 
security:
  - bearerAuth: []

paths:


  # AUTH
  
  /users:
    post:
      security: []
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input

  /auth/login:
    post:
      security: []
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized


  # PUBLIC DATA
  /horses:
    get:
      summary: Get all horses
      responses:
        '200':
          description: List of horses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Horse'

  /trainers:
    get:
      summary: Get all trainers
      responses:
        '200':
          description: List of trainers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trainer'

  /training-types:
    get:
      summary: Get all training types
      responses:
        '200':
          description: List of training types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingType'


  # TRAININGS

  /trainings:
    get:
      summary: Get all trainings
      responses:
        '200':
          description: List of trainings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Training'

    post:
      summary: Create training slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingCreate'
      responses:
        '201':
          description: Training created
        '400':
          description: Validation error

  /trainings/page:
    get:
      summary: Get paginated and filtered trainings
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, full, completed, cancelled]
        - name: horse_id
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Paginated trainings matching the filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingPage'

  /trainings/{id}:
    put:
      summary: Update training details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingUpdate'
      responses:
        '200':
          description: Training successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
        '400':
          description: Validation error / Invalid input
        '404':
          description: Training not found

    delete:
      summary: Delete training
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Training not found

  /trainings/{id}/join:
    post:
      summary: Join training (book a spot)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinTraining'
      responses:
        '201':
          description: Successfully joined
        '400':
          description: Training is full or invalid request
        '404':
          description: Training not found

  /trainings/{id}/participants:
    get:
      summary: Get participants of training
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingParticipant'


# COMPONENTS

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    TrainingUpdate:
      type: object
      properties:
        horse_id:
          type: integer
        trainer_id:
          type: integer
        training_type_id:
          type: integer
        date:
          type: string
          format: date-time
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, completed, cancelled]

    UserRegister:
      type: object
      required: [name, surname, email, password]
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string

    Horse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        discipline:
          type: string
        daily_limit:
          type: integer

    Trainer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        specialization:
          type: string

    TrainingType:
      type: object
      properties:
        id:
          type: integer
        discipline:
          type: string
          enum: [Dressage, Jumping, Recreational]
        training_mode:
          type: string
          enum: [individual, group]

    Training:
      type: object
      properties:
        id:
          type: integer
        horse_id:
          type: integer
        trainer_id:
          type: integer
        training_type_id:
          type: integer
        date:
          type: string
          format: date-time
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, completed, cancelled]
        created_at:
          type: string
          format: date-time

    TrainingPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Training'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer

    TrainingCreate:
      type: object
      required: [horse_id, trainer_id, training_type_id, date]
      properties:
        horse_id:
          type: integer
        trainer_id:
          type: integer
        training_type_id:
          type: integer
        date:
          type: string
          format: date-time

    JoinTraining:
      type: object
      properties: {}
        

    TrainingParticipant:
      type: object
      properties:
        id:
          type: integer
        training_id:
          type: integer
        user_id:
          type: integer